Adding the ability to migrate from existing in db views to in memory views.
